feat: align with marks
